Rename Folder/Class:  psr-4 convention (Using an prefix class name same as container folder) more convenience using namespaces (omit of uses the require_once files)
