+++ includes/Obiba/ObibaMicaClient/MicaClient/DrupalMicaClient/DrupalMicaClientTempFileResource.php
@@ -13,7 +13,7 @@
  * @file
  * TempFile resource class used to communicate with backend server
  */
-namespace Obiba\ObibaMicaClient\MicaClient\DrupalMicaClientClasses;
+namespace Obiba\ObibaMicaClient\MicaClient\DrupalMicaClient;
 
 use Obiba\ObibaMicaClient\MicaCache as MicaCache;
 use Obiba\ObibaMicaClient\MicaConfigurations as MicaConfig;
