+++ obiba_mica_data_access_request/obiba_mica_data_access_request.module
@@ -9,7 +9,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-use Obiba\ObibaMicaClient\MicaClient\DrupalMicaClientClasses as DrupalMicaClientClasses;
+use Obiba\ObibaMicaClient\MicaClient\DrupalMicaClient as DrupalMicaClient;
 
 /**
  * Implements hook_menu().
@@ -317,7 +317,7 @@ and intellectual property policies.</p> </div>";
  * Data Access Request Title Callback
  */
 function obiba_mica_data_access_request_dar_title_callback() {
-  $permissionResource = new DrupalMicaClientClasses\DrupalMicaClientPermission();
+  $permissionResource = new DrupalMicaClient\DrupalMicaClientPermission();
   return $permissionResource->userHasRole('mica-data-access-officer') ?
     variable_get_value('access_requests_page_title') :
     variable_get_value('access_request_my_page_title');
